Request to update odd_even.py
This is a python program. Start using pythonic way of coding. The main emphasis on using python to code, is to shorten the coding time. Hence try to make it simple. I didn't get any data from the user as I asked in the question, that's not mandatory. Have fun coding!